+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Change Log
 ## Unreleased
 * Promote `typing-extensions` to runtime dependency
 * Improved disk write performance when encoding items
+* You can now hook into collection updates and run your own logic using the
+  `alternatives.update_item` event.
 
 ## v0.13.3 - 2025-05-12
 * Add support for beets v2.3.x